Sanitize directory structure

deleted:    thoth/lab/viz/js/dynamic/__init__.py
deleted:    thoth/lab/viz/js/static/__init__.py
modified:   setup.py
renamed:    thoth/lab/viz/js/__init__.py -> thoth/lab/viz/__init__.py
renamed:    thoth/lab/viz/js/dynamic/templates/diagonal.js -> thoth/lab/viz/assets/js/dynamic/templates/diagonal.js
renamed:    thoth/lab/viz/js/static/templates/diagonal.js -> thoth/lab/viz/assets/js/static/templates/diagonal.js
