+++ thoth/lab/viz/assets/js/dynamic/templates/diagonal.js
