+++ thoth/lab/viz/assets/js/static/templates/diagonal.js
