--- thoth/lab/viz/js/dynamic/__init__.py
@@ -1,1 +1,0 @@
-"""Dynamic visualization layouts."""
