--- thoth/lab/viz/js/static/__init__.py
@@ -1,1 +1,0 @@
-"""Static visualization layouts."""
